Restrict users from adding duplicate songs.
Update PostSong task in Server/Controllers/SongController.cs to check for duplicates for the Song attempting to be added. If a song already exists with the same Title, Artist, Album, and Genre, the Song is not added and a Conflict response is returned

Update HandleValidSubmit Task in Client/Pages/AddSong.razor. Checks to see if the http response has a conflict status code; if it does, 'message' gets set, informing the user of duplicate songs